+++ libsolidity/formal/SymbolicState.h
@@ -189,13 +189,13 @@ private:
 	BlockchainVariable m_tx{
 		"tx",
 		{
-			{"blockhash", std::make_shared<smtutil::ArraySort>(smtutil::SortProvider::uintSort, smtutil::SortProvider::uintSort)},
 			{"block.chainid", smtutil::SortProvider::uintSort},
 			{"block.coinbase", smt::smtSort(*TypeProvider::address())},
 			{"block.difficulty", smtutil::SortProvider::uintSort},
 			{"block.gaslimit", smtutil::SortProvider::uintSort},
 			{"block.number", smtutil::SortProvider::uintSort},
 			{"block.timestamp", smtutil::SortProvider::uintSort},
+			{"blockhash", std::make_shared<smtutil::ArraySort>(smtutil::SortProvider::uintSort, smtutil::SortProvider::uintSort)},
 			// TODO gasleft
 			{"msg.data", smt::smtSort(*TypeProvider::bytesMemory())},
 			{"msg.sender", smt::smtSort(*TypeProvider::address())},