+++ test/libsolidity/smtCheckerTests/special/msg_value_3.sol
@@ -1,0 +1,13 @@
+pragma experimental SMTChecker;
+contract C {
+	bool lock = true;
+	function f() public {
+		lock = false;
+		g();
+		lock = true;
+	}
+	function g() public payable {
+		require(lock == false);
+		assert(msg.value == 0);
+	}
+}