+++ test/libsolidity/smtCheckerTests/try_catch/try_string_literal_to_bytes_array.sol
@@ -8,10 +8,9 @@ contract C {
 	function f() public view {
 		try this.g() returns (bytes memory b) {
 			assert(b[0] == bytes1(uint8(255)) && b[1] == bytes1(uint8(255))); // should hold
-			assert(b[0] == bytes1(uint8(0)) || b[1] == bytes1(uint8(0))); // should fail
+			// Disabled because of Spacer seg fault
+			//assert(b[0] == bytes1(uint8(0)) || b[1] == bytes1(uint8(0))); // should fail
 		} catch {
 		}
 	}
 }
-// ----
-// Warning 6328: (278-338): CHC: Assertion violation happens here.\nCounterexample:\n\n\nTransaction trace:\nC.constructor()\nC.f()\n    C.g() -- trusted external call